fix(client): preserve recovery fallback and clarify cookie jar states
