+++ docs/docs/builders/packages/client-pwa-token-audit.generated.md
@@ -121,13 +121,13 @@ Approved contrast exceptions:
 | `packages/client/src/components/Public/atoms/EditorialSelect.tsx:91` | `text-primary-action` | action | <code>className="absolute right-3 inline-flex h-4 w-4 items-center justify-center text-xs text-primary-action"</code> |
 | `packages/client/src/components/Public/PublicActionCard.tsx:41` | `ring-primary-action` | action | <code>className="group flex h-full flex-col gap-4 text-left focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary-action focus-visible:ring-offset-2"</code> |
 | `packages/client/src/components/Public/PublicActionCard.tsx:59` | `text-primary-action` | action | <code>className="font-serif text-xl font-normal leading-[1.18] tracking-[-0.01em] text-text-strong-950 group-hover:text-primary-action"</code> |
-| `packages/client/src/components/Public/PublicCookieJarCard.tsx:314` | `ring-primary-action` | action | <code>"flex h-full flex-col gap-4 border-t border-stroke-soft-200 pt-5 text-left focus:outline-none focus-visible:ring-2 focus-visible:ring-primary-action focus-visible:ring-offset-2",</code> |
-| `packages/client/src/components/Public/PublicCookieJarCard.tsx:315` | `ring-primary-action` | action | <code>isHighlighted &amp;&amp; "ring-2 ring-primary-action ring-offset-2"</code> |
-| `packages/client/src/components/Public/PublicCookieJarCard.tsx:406` | `text-primary-action` | action | <code>className="text-text-sub-600 underline decoration-stroke-soft-200 underline-offset-4 transition hover:text-primary-action"</code> |
-| `packages/client/src/components/Public/PublicCookieJarCard.tsx:631` | `border-primary-base` | state | <code>inputClassName="mt-2 w-full rounded-lg border border-stroke-soft-200 bg-bg-white-0 px-3 py-2 text-sm text-text-strong-950 outline-none focus:border-primary-base focus:ring-2 focus:ring-primary-base/30"</code> |
-| `packages/client/src/components/Public/PublicCookieJarCard.tsx:631` | `ring-primary-base` | state | <code>inputClassName="mt-2 w-full rounded-lg border border-stroke-soft-200 bg-bg-white-0 px-3 py-2 text-sm text-text-strong-950 outline-none focus:border-primary-base focus:ring-2 focus:ring-primary-base/30"</code> |
-| `packages/client/src/components/Public/PublicCookieJarCard.tsx:681` | `border-primary-base` | state | <code>inputClassName="mt-2 w-full rounded-lg border border-stroke-soft-200 bg-bg-white-0 px-3 py-2 text-sm text-text-strong-950 outline-none focus:border-primary-base focus:ring-2 focus:ring-primary-base/30"</code> |
-| `packages/client/src/components/Public/PublicCookieJarCard.tsx:681` | `ring-primary-base` | state | <code>inputClassName="mt-2 w-full rounded-lg border border-stroke-soft-200 bg-bg-white-0 px-3 py-2 text-sm text-text-strong-950 outline-none focus:border-primary-base focus:ring-2 focus:ring-primary-base/30"</code> |
+| `packages/client/src/components/Public/PublicCookieJarCard.tsx:276` | `ring-primary-action` | action | <code>"flex h-full flex-col gap-4 border-t border-stroke-soft-200 pt-5 text-left focus:outline-none focus-visible:ring-2 focus-visible:ring-primary-action focus-visible:ring-offset-2",</code> |
+| `packages/client/src/components/Public/PublicCookieJarCard.tsx:277` | `ring-primary-action` | action | <code>isHighlighted &amp;&amp; "ring-2 ring-primary-action ring-offset-2"</code> |
+| `packages/client/src/components/Public/PublicCookieJarCard.tsx:368` | `text-primary-action` | action | <code>className="text-text-sub-600 underline decoration-stroke-soft-200 underline-offset-4 transition hover:text-primary-action"</code> |
+| `packages/client/src/components/Public/PublicCookieJarCard.tsx:594` | `border-primary-base` | state | <code>inputClassName="mt-2 w-full rounded-lg border border-stroke-soft-200 bg-bg-white-0 px-3 py-2 text-sm text-text-strong-950 outline-none focus:border-primary-base focus:ring-2 focus:ring-primary-base/30"</code> |
+| `packages/client/src/components/Public/PublicCookieJarCard.tsx:594` | `ring-primary-base` | state | <code>inputClassName="mt-2 w-full rounded-lg border border-stroke-soft-200 bg-bg-white-0 px-3 py-2 text-sm text-text-strong-950 outline-none focus:border-primary-base focus:ring-2 focus:ring-primary-base/30"</code> |
+| `packages/client/src/components/Public/PublicCookieJarCard.tsx:644` | `border-primary-base` | state | <code>inputClassName="mt-2 w-full rounded-lg border border-stroke-soft-200 bg-bg-white-0 px-3 py-2 text-sm text-text-strong-950 outline-none focus:border-primary-base focus:ring-2 focus:ring-primary-base/30"</code> |
+| `packages/client/src/components/Public/PublicCookieJarCard.tsx:644` | `ring-primary-base` | state | <code>inputClassName="mt-2 w-full rounded-lg border border-stroke-soft-200 bg-bg-white-0 px-3 py-2 text-sm text-text-strong-950 outline-none focus:border-primary-base focus:ring-2 focus:ring-primary-base/30"</code> |
 | `packages/client/src/components/Public/PublicEndowmentPanel.tsx:201` | `ring-primary-action` | action | <code>className="inline-flex h-10 w-10 shrink-0 cursor-pointer items-center justify-center rounded-full border border-stroke-soft-200 bg-bg-white-0 text-text-sub-600 transition-colors hover:bg-bg-weak-50 hover:text-text-strong-950 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary-action focus-visible:ring-offset-2"</code> |
 | `packages/client/src/components/Public/PublicEndowmentPanel.tsx:636` | `border-primary-action` | action | <code>"min-h-11 w-full rounded-full border bg-bg-white-0 px-4 py-2.5 text-sm text-text-strong-950 outline-none transition-colors placeholder:text-text-soft-400 focus:border-primary-action",</code> |
 | `packages/client/src/components/Public/PublicEvidenceCard.tsx:243` | `ring-primary-action` | action | <code>className="group flex h-full cursor-pointer flex-col gap-3 text-left focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary-action focus-visible:ring-offset-2"</code> |
@@ -256,7 +256,7 @@ Approved contrast exceptions:
 | `packages/client/src/components/Pwa/statusStyles.ts:27` | `ring-primary-alpha-24` | state | <code>"focus-visible:ring-2 focus-visible:ring-primary-alpha-24 focus-visible:border-primary active:border-primary",</code> |
 | `packages/client/src/components/Pwa/statusStyles.ts:28` | `text-primary-accent-foreground` | accent | <code>foreground: "text-primary-accent-foreground",</code> |
 | `packages/client/src/index.css:93` | `primary-base` | accent | <code>background: rgb(var(--tone-primary, var(--primary-base)) / 0.32);</code> |
-| `packages/client/src/styles/editorial.css:560` | `primary-action` | action | <code>outline: 2px solid var(--color-primary-action);</code> |
+| `packages/client/src/styles/editorial.css:604` | `primary-action` | action | <code>outline: 2px solid var(--color-primary-action);</code> |
 | `packages/client/src/views/Garden/Details.tsx:245` | `bg-primary-base` | contrast-exception | <code>? "bg-primary-base text-primary-accent-foreground border-primary-base"</code> |
 | `packages/client/src/views/Garden/Details.tsx:245` | `border-primary-base` | contrast-exception | <code>? "bg-primary-base text-primary-accent-foreground border-primary-base"</code> |
 | `packages/client/src/views/Garden/Details.tsx:245` | `text-primary-accent-foreground` | contrast-exception | <code>? "bg-primary-base text-primary-accent-foreground border-primary-base"</code> |
