+++ packages/client/src/components/Public/PublicSurfaceState.tsx
@@ -1,5 +1,6 @@
 import type { PublicSurfaceState as PublicSurfaceStateValue } from "@green-goods/shared/public";
 import type { ReactNode } from "react";
+import { FormattedMessage } from "react-intl";
 
 export interface PublicSurfaceStateProps {
   state: PublicSurfaceStateValue;
@@ -23,6 +24,9 @@ export function PublicSurfaceState({
   if (state === "loading") {
     return (
       <Container aria-busy="true" data-public-surface-state={state}>
+        <span className="sr-only">
+          <FormattedMessage id="app.common.loading" defaultMessage="Loading..." />
+        </span>
         {loading}
       </Container>
     );
